+++ src/libmetis/auxapi.rs
@@ -14,6 +14,6 @@ pub unsafe extern "C" fn METIS_Free(mut ptr: *mut libc::c_void) -> libc::c_int {
 }
 #[no_mangle]
 pub unsafe extern "C" fn METIS_SetDefaultOptions(mut options: *mut idx_t) -> libc::c_int {
-    libmetis__iset(40 as libc::c_int as size_t, -(1 as libc::c_int), options);
+    libmetis__iset(40 as libc::c_int as size_t, -(1), options);
     return METIS_OK as libc::c_int;
 }
